Make input pixels in the range 0-1 for NVVFX SDK compatibility

+++ Plugins/NVVFX/Source/CommonCUDAKernels.h
@@ -4,5 +4,5 @@
 #include "device_launch_parameters.h"
 
 extern "C" {
-	void NormalizeKernelWrapper(dim3 blocks, dim3 thread, int* data, int max, int dataSize);
+	void NormalizeKernelWrapper(dim3 blocks, dim3 thread, float* data, float max, int dataSize);
 }
